Refactor Chevron tests
