+++ .lintstagedrc.json
@@ -1,3 +1,5 @@
 {
+  "*.{js,mjs,ts,tsx,mdx}": "eslint --fix",
+  "*.{css,scss}": "stylelint --fix",
   "*.{json,jsonc,md,yml,xml,svg}": "prettier --write"
 }
