--- packages/.lintstagedrc.json
@@ -1,5 +1,0 @@
-{
-  "*.{js,mjs,ts,tsx,mdx}": "eslint --fix",
-  "*.{css,scss}": "stylelint --fix",
-  "*.{json,jsonc,md,yml,xml,svg}": "prettier --write"
-}
