+++ packages/fuselage/src/components/Chevron/Chevron.spec.tsx
@@ -1,10 +1,8 @@
+import { render } from '@testing-library/react';
 import React from 'react';
-import ReactDOM from 'react-dom';
 
-import { Chevron } from '../..';
+import { Chevron } from '.';
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<Chevron />, div);
-  ReactDOM.unmountComponentAtNode(div);
+  render(<Chevron />);
 });
