+++ packages/fuselage/src/components/Divider/Divider.spec.tsx
@@ -1,10 +1,8 @@
+import { render } from '@testing-library/react';
 import React from 'react';
-import ReactDOM from 'react-dom';
 
-import { Divider } from '../..';
+import { Divider } from '.';
 
 it('renders without crashing', () => {
-  const div = document.createElement('div');
-  ReactDOM.render(<Divider />, div);
-  ReactDOM.unmountComponentAtNode(div);
+  render(<Divider />);
 });
